erreurs

+++ Coccinelles_2025/superstar_asserv/include/Ultrason.hpp
@@ -16,11 +16,11 @@ class Ultrason
 private:
     int m_echo_PIN;     // Broche Echo du HC-SR04
     int m_trigger_PIN;  // Broche Trigger du HC-SR04
-    long m_duration; // Durée de l'echo
+    double m_duration; // Durée de l'echo
     long m_dt = 10; // Délai entre les lectures (10ms)
 public:
-    long m_distance; // Distance mesurée
-    long m_time;
+    double m_distance; // Distance mesurée
+    double m_time;
     Ultrason(int m_echo_PIN, int m_trigger_PIN); // Constructeur
     /**
      * @brief Fonction d'initialisation du capteur ultrason
